% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseSeqs}
\alias{baseSeqs}
\title{Wild type CYP isoform sequences}
\format{
{
Named vector of CYP wild sequences
\describe{
\item{name}{Name of CYP isoform.}
\item{baseSeq}{aminoacid sequence of wild type isoform}
}
}
}
\source{
UniprotKB (2021).
}
\usage{
baseSeqs
}
\description{
Wild type CYP isoform sequences
}
\examples{
\dontrun{
 baseSeqs
}
}
\references{
The UniProt Consortium
UniProt: the universal protein knowledgeable in 2021
Nucleic Acids Res. 49:D1 (2021)
}
\keyword{datasets}
