% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drugs}
\alias{drugs}
\title{In vitro and in vivo current validated CYP - drug interactions}
\format{
{
Named vector of CYP drug interaction:
\describe{
\item{name}{CYP isoform}
\item{drugs}{List of drugs interacting with CYP isoform.}
}
}
}
\source{
FDA, drug interaction and labeling (2021).
}
\usage{
drugs
}
\description{
In vitro and in vivo current validated CYP - drug interactions
}
\examples{
\dontrun{
 drugs
}
}
\references{
Ministry of Health, Labour and Welfare (MHLW), Japan (2014). Drug interaction guideline for drug development and labeling recommendations
European Medicines Agency (2013). Guideline on the Investigation of Drug Interactions.
}
\keyword{datasets}
