% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsnpCYP.R
\name{runsnpCYP}
\alias{runsnpCYP}
\title{This will launch Shiny implementation of snpCYP}
\usage{
runsnpCYP()
}
\value{
No return value, should open up a Shiny webpage.
}
\description{
A function to run Shiny application for snpCYP
}
\examples{
\dontrun{
snpCYP::runsnpCYP()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
    \href{https://shiny.rstudio.com/tutorial/}{Link}
}
