% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpDist.R
\name{snpDist}
\alias{snpDist}
\title{Generating a graphic output of the frequency of nsSNPs for each region
of the provided CYP isoform protein.}
\usage{
snpDist(CYP, seq)
}
\arguments{
\item{CYP}{CYP isoform to be analyzed, string}

\item{seq}{Aminoacid sequence (one-letter code) for the CYP isoform to be
analyzed (list of strings)}
}
\value{
Visualization of the frequency of nsSNPs for each protein region
         Also return the SNPs positions for each region
}
\description{
This function generates a graphic output of the frequency of nsSNPs for
each sequential group of 50 aminoacids at the CYP isoform provided.
}
\examples{
\dontrun{
snpDist("CYP1A2",
    paste("MALSQSVPFSATELLLASAIFCLVFWVLKGLRPRVPKGLKSPPEPWGWPLLGHVLTLGKN",
          "PHLALSRMSQRYGDVLQIRIGSTPVLVLSRLDTIRQALVRQGDDFKGRPDLYTSTLITDG",
          "QSLTFSTDSGPVWAARRRLAQNALNTFSIASDPASSSSCYLEEHVSKEAKALISRLQELM",
          "AGPGHFDPYNQVVVSVANVIGAMCFGQHFPESSDEMLSLVKNTHEFVETASSGNPLDFFP",
          "ILRYLPNPALQRFKAFNQRFLWFLQKTVQEHYQDFDKNSVRDITGALFKHSKKGPRASGN",
          "LIPQEKIVNLVNDIFGAGFDTVTTAISWSLMYLVTKPEIQRKIQKELDTVIGRERRPRLS",
          "DRPQLPYLEAFILETFRHSSFLPFTIPHSTTRDTTLNGFYIPKKCCVFVNQWQVNHDPEL",
          "WEDPSEFRPERFLTADGTAINKPLSEKMMLFGMGKRRCIGEVLAKWEIFLFLAILLQQLE",
          "FSVPPGVKVDLTPIYGLTMKHARCEHVQARLRFSIN", sep=""))
}


}
