% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpToPred.R
\name{snpToPred}
\alias{snpToPred}
\title{Generating a prediction of the potential phenotypical damage a snp may cause}
\usage{
snpToPred(prot, pos, aa1, aa2)
}
\arguments{
\item{prot}{UniProtKB identifier of the CYP to be analyzed, string}

\item{pos}{Position of the nucleotide polymorphism, string}

\item{aa1}{Original (or wild type) aminoacid, single letter code, char}

\item{aa2}{Mutant aminoacid, single letter code, char}
}
\value{
Probability of the provided mutation to cause phenotypical
enzymatic damage.
}
\description{
This function generates a probabilistic prediction of damage by probing
Polyphen-2 server engine and databases, which in turn perform analysis
on the sequence, phylogenetic and structural information related to the
provided substitution.
}
\details{
The function also returns the false positive rate and true positive rate
associated with the prediction
}
\examples{
\dontrun{
snpToPred("O95479", "453", "R", "Q")
}

}
\references{
Adzhubei, I., Jordan, D. M., & Sunyaev, S. R. (2013).
Predicting functional effect of human missense mutations using PolyPhen-2.
Current Protocols in Human Genetics, 76 (1), 7–20.
}
